business_analytics/requirements.txt
django>=4.2.0
django-rest-framework>=0.1.0
pandas>=1.5.0
numpy>=1.22.0
openpyxl>=3.1.0
google-generativeai>=0.3.0
matplotlib>=3.7.0
scikit-learn>=1.0.0
python-dotenv>=1.0.0 